modify module of student

+++ yunzaodao-后端/eladmin-master/eladmin-system/src/main/java/com/yzd/modules/system/repository/CourseRepository.java
@@ -11,7 +11,7 @@ import java.util.List;
 * @author wdc
 * @date 2020-03-20
 */
-public interface CourseRepository extends JpaRepository<Course, Integer>, JpaSpecificationExecutor<Course> {
+public interface CourseRepository extends JpaRepository<Course, Long>, JpaSpecificationExecutor<Course> {
 
     @Query(value = "select course_code from course" , nativeQuery = true)
     List<String> findCourseCodes();
